Slow down tick rate for Windows 10 2004 users
Scheduler behavior changed on Win10 2004, and has a default resolution of best-approximation-possible-for-15.625ms. In order to accommodate this lower-resolution scheduler, we have to slow down the tick rate.

The correct solution would be to call `timeBeginPeriod(1)` on Windows 10 only, but doing this for each language in a short and portable way is hard.